requirements.txt
Abg==2.3.8
aiofiles
aiohttp
httpx[http2]>=0.18.1,<1.0.0
alphabet_detector==0.0.7
beautifulsoup4==4.11.1
bleach==6.1.0
cachetools
configparser==5.3.0
dateparser
emoji==1.7.0
feedparser==6.0.10
ffmpeg-python==0.2.0
future==0.18.3
gitpython==3.1.43
markdown2==2.4.13
multicolorcaptcha==1.2.0
Pillow>=10.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyYAML==6.0.1
pyrate-limiter==2.8.5
python-telegram-bot==13.15
requests==2.31.0
spamwatch==0.3.0
speedtest-cli==2.1.3
SQLAlchemy==1.4.44
telethon==1.35.0
ujson==5.8.0
pycryptodome==3.20.0
prettyconf
base58==2.1.1
pytimeparse==1.1.8
SibylSystem==0.0.24
TgCrypto
Pyrofork==2.3.21.post3
python-arq==6.0.7
python-dotenv==1.0.1
ffmpeg-python==0.2.0
redis
telegraph==2.2.0
motor
pymongo[srv]==4.7.2
odmantic==0.9.2
numpy<2.0.0
opencv-python-headless
pykeyboard==0.1.5
wget==3.2
wikipedia==1.4.0
bing_image_downloader==1.1.2
lyricsgenius==3.0.1
jikanpy==4.3.2
gtts==2.5.1
search_engine_parser==0.6.8
googletrans==2.4.0
gpytranslate==1.5.1
cloudscraper==1.2.70
carbonnow
pretty_errors
